+++ source/renderer/app/components/wallet/WalletSendConfirmationDialog.js
@@ -103,7 +103,7 @@ type Props = {
   onCancel: Function,
   onExternalLinkClick: Function,
   isSubmitting: boolean,
-  isMainnet: boolean,
+  isFlight: boolean,
   error: ?LocalizableError,
   currencyUnit: string,
 };
@@ -183,7 +183,7 @@ export default class WalletSendConfirmationDialog extends Component<Props> {
       totalAmount,
       transactionFee,
       isSubmitting,
-      isMainnet,
+      isFlight,
       error,
       currencyUnit,
       onExternalLinkClick,
@@ -206,7 +206,7 @@ export default class WalletSendConfirmationDialog extends Component<Props> {
         className: confirmButtonClasses,
         disabled:
           !passphraseField.isValid ||
-          (!flightCandidateCheckboxField.value && isMainnet),
+          (!flightCandidateCheckboxField.value && isFlight),
       },
     ];
 
@@ -290,7 +290,7 @@ export default class WalletSendConfirmationDialog extends Component<Props> {
           />
         </div>
 
-        {isMainnet && (
+        {isFlight && (
           <div className={styles.flightCandidateWarning}>
             <FormattedHTMLMessage
               {...messages.flightCandidateWarning}